Lmr tune (#15)
* Parametrized UCI options

* Reworked LMR, this is initial partial tune

* Updated version number
Bench 1079851

ELO      | 6.20 +- 4.37 (95%)
SPRT   | 60.0+0.6s Threads=1 Hash=64MB
LLR      | 2.97 (-2.94, 2.94) [0.00, 5.00]
Games | N: 9808 W: 2075 L: 1900 D: 5833

ELO     | 18.63 +- 9.08 (95%)
SPRT   | 12.0+0.12s Threads=1 Hash=8MB
LLR      | 2.99 (-2.94, 2.94) [0.00, 5.00]
Games | N: 2912 W: 831 L: 675 D: 1406

+++ src/transposition.h
@@ -69,7 +69,7 @@ public:
                          // 16 bytes total, 4 entries per cache line.
 };
 
-void SetTableSize(const size_t MB);
+void SetTableSize(const int32_t MB);
 Entry GetEntry(const HashType hash);
 void SaveEntry(const Board &board, const Move best_move, const Score score,
                const Depth depth, const uint8_t bound = kLowerBound);
